Deleted Deck class and moved code to DeckParser
