+++ .idea/.name
@@ -1,1 +1,1 @@
-TopTrumps
+MScIT2020Project
